Fix for ptree.c addUnparsedField() string termination issue
The addUnparsedField() was not passing the string lengths and relying on null termination.
This can be a problem when ln_normalize() is fed an with an unterminated string as the input;
When an input parses correctly there is no issue, only unparsed inputs will trigger this issue.
To duplicate the bug, simply pass a string to ln_normalize that is known to fail logparsing,
but pass in a shorter string length. The output will contain the entire string regardless.