API: error messages also sent to debug log
also one more error message shuffled to new interface. That's it for
now in regard to error messages, but we will review the code and add
more. The important part is that callers can begin to use the new
API and thus will benefit when we add more error messages.