add runtime path information
Note: this is PoC code, NOT to be used in practice. We will
evolve this to a real solution with the next commits. We do
not squash them, because there are a couple of interim steps
which I would like to see documented.

The reason is that we must make some architectural changes
to the parser and other interfaces in order to support the
runtime path information option as a permanent feature. These
architecture changes shall be clearly visisble from the git log
and, while releated, are definitely for a larger goal than what
we currently have.