add support for upcoming new versions of json-c
The new API json_global_set_stringbuf_initial_size() permits us
to increase the far too low default buffer size. This results in
radically fewer malloc calls and thus performance enhancements.

Note that this is compatible with previous versions of json-c,
as we check if the new API is available.