Use Libs.private instead of Requires.private in pkgconfig
Requires.private has an unfortunate side effect: when pkg-config
--cflags is called, if the pkg-config of a Requires.private library
is not available, it will fail. This is very counter-intuitive, as
one would expect Requires.private to be parsed only if pkg-config is
called with --static, but that's the case only for --libs, not
--cflags.

Autogenerate Libs.private from libestr pkg-config --libs output via
autoconf.