FIX: CURSOR OUT OF BOUNDS WHEN LAST VALUE IS EMPTY
- Expected outcome:
When the last value is empty, the cursor should stop at the equal sign.

- Problem encountered:
When the last value is empty, the cursor increases (+2): i > npb->strlen

- Fix:
Initialized the length of the value to 0: lenValue = 0
If the cursor has not reached the end of the line, we parse the value.
Otherwise, we know that our value is empty, its length is thus 0.